SecureQRLoginSystem/requirements.txt
opencv-python-headless
qrcode[pil]
Pillow
tkcalendar